+++ state/sqlserver/v1/metadata.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package sqlserver
+package v1
 
 import (
 	"encoding/json"
