+++ state/sqlserver/v1/sqlserver_test.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package sqlserver
+package v1
 
 import (
 	"context"
