+++ tests/certification/state/sqlserver/v1/components/azure/localsecrets.yaml
