+++ tests/certification/state/sqlserver/v1/components/azure/sqlserver.yaml
