+++ tests/certification/state/sqlserver/v1/components/docker/customschemawithindex/sqlserver.yaml
