+++ tests/certification/state/sqlserver/v1/components/docker/default/sqlserver.yaml
